+++ CHANGELOG.rst
@@ -79,6 +79,7 @@ Fixed
 - Fixed :issue:`534` - flipped axes in NSA scatterplot when plotting absmag colors
 - Fixed :issue:`559` - bug in check_marvin when marvindb is None
 - Fixed :issue:`579` - bug in MMA with marvindb preventing files from opening
+- Fixed :issue:`543`, :issue:`552`, :issue:`553` - bugs with various Query handlings
 
 Refactored
 ^^^^^^^^^^